Add server-side, update app architecture & code improvements
