+++ bbb-self-test/.gitignore
@@ -1,9 +1,1 @@
-.dbshell
-npm-debug.log
-node_modules/
-.meteor/dev_bundle
-public/locales/de_DE.json
-public/locales/ja_JP.json
-public/files
-
-
+.idea
