+++ bbb-self-test/client/README.md
