+++ bbb-self-test/client/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "bbb-self-test",
-  "description": "BigBlueButton Self Test Application",
+  "name": "bbb-self-test-client",
+  "description": "BigBlueButton Self Test Application Frontend",
   "version": "0.1.0",
   "license": "LGPL-3.0",
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
-    "eject": "react-app-rewired eject"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "dependencies": {
     "@ant-design/icons": "^5.0.1",
@@ -19,30 +19,14 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "antd": "^5.3.0",
-    "async_hooks": "^1.0.0",
-    "browserify-zlib": "^0.2.0",
-    "buffer": "^6.0.3",
-    "child_process": "^1.0.2",
-    "crypto-browserify": "^3.12.0",
+    "axios": "^1.3.4",
     "dayjs": "^1.11.7",
     "detectrtc": "^1.4.1",
-    "fs": "^0.0.1-security",
-    "http": "^0.0.1-security",
-    "https-browserify": "^1.0.0",
-    "net": "^1.0.2",
-    "network-speed": "^2.1.1",
-    "node-polyfill-webpack-plugin": "^2.0.1",
     "react": "^18.2.0",
-    "react-app-rewired": "^2.2.1",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "speedtest-net": "^2.2.0",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "tls": "^0.0.1",
     "turn-it-off": "^1.0.6",
     "typescript": "^4.9.5",
-    "universal-speedtest": "^2.0.6",
     "web-vitals": "^2.1.4"
   },
   "devDependencies": {
@@ -51,9 +35,7 @@
     "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
-    "eslint-webpack-plugin": "^4.0.0",
-    "process": "^0.11.10",
-    "webpack": "^5.0.0"
+    "eslint-webpack-plugin": "^4.0.0"
   },
   "eslintConfig": {
     "extends": [
