+++ bbb-self-test/client/public/bg.jpg
