+++ bbb-self-test/client/public/favicon.ico
