+++ bbb-self-test/client/public/index.html
