+++ bbb-self-test/client/public/logo192.png
