+++ bbb-self-test/client/public/logo512.png
