+++ bbb-self-test/client/public/manifest.json
