+++ bbb-self-test/client/src/App.test.tsx
