+++ bbb-self-test/client/src/functions/delay.function.ts
@@ -1,0 +1,1 @@
+export const delay = ms => new Promise(res => setTimeout(res, ms));
