+++ bbb-self-test/client/src/functions/downloadJson.function.ts
@@ -1,0 +1,6 @@
+export const downloadJson = (jsonObject: object) => {
+    const downloadNode = document.createElement('a');
+    downloadNode.download = "bbb-checks.json";
+    downloadNode.href =  "data:text/json;charset=utf-8," + encodeURIComponent(JSON.stringify(jsonObject));
+    downloadNode.click();
+}
