+++ bbb-self-test/client/src/index.tsx
