+++ bbb-self-test/client/src/react-app-env.d.ts
