+++ bbb-self-test/client/src/services/network-speed.service.ts
@@ -1,0 +1,14 @@
+import axios from 'axios';
+import { apiRoutes } from '../routing/backend-config';
+
+class NetworkSpeedService {
+    getDownloadSpeed() {
+        return axios.get(apiRoutes.DOWNLOAD_URL);
+    }
+
+    getUploadSpeed() {
+        return axios.get(apiRoutes.UPLOAD_URL);
+    }
+}
+
+export default new NetworkSpeedService();
