+++ bbb-self-test/client/src/setupTests.ts
