+++ bbb-self-test/client/src/types/NetworkType.ts
