+++ bbb-self-test/client/tsconfig.json
