+++ bbb-self-test/server/.gitignore
@@ -1,0 +1,1 @@
+node_modules/
